requirements.txt
streamlit>=1.40,<2
requests>=2.32,<3
pandas>=2.2,<3
python-dateutil>=2.9,<3
psycopg2-binary>=2.9,<3
